pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memos-mcp"
version = "0.1.0"
description = "A streamable HTTP MCP server for Memos note-taking app"
authors = [
    {name = "Ivan Lee", email = "ivan@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/ivanlee1999/memos-mcp"
Repository = "https://github.com/ivanlee1999/memos-mcp"
Issues = "https://github.com/ivanlee1999/memos-mcp/issues"

[project.scripts]
memos-mcp = "memos_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/memos_mcp"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true